requirements.txt
fastapi #==0.119.0
httpx #== 0.28.1
psycopg2-binary #==2.9.11
pydantic #==2.12.2
pydantic-settings #==2.11.0
python-dotenv #==1.1.1
SQLAlchemy #==2.0.44
requests #==2.32.5
starlette #==0.48.0
uvicorn #==0.37.0
typing_inspect #==0.9.0